requirement.txt
certifi==2024.7.4   
charset-normalizer==3.3.2      
contourpy==1.2.1      
cramjam==2.8.3      
cycler==0.12.1     
distlib==0.3.8      
fastparquet==2024.5.0   
filelock==3.15.4     
fonttools==4.53.1     
fsspec==2024.6.1   
greenlet==3.0.3
idna==3.7
kiwisolver==1.4.5
matplotlib==3.9.1
mysql-connector-python==9.0.0
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pip==24.1.2
platformdirs==4.2.2
py4j==0.10.9.7
pyarrow==16.1.0
PyMySQL==1.1.1
pyparsing==3.1.2
pyspark==3.5.1
python-dateutil==2.9.0.post0
pytz==2024.1
requests==2.32.3
retrying==1.3.4
setuptools==65.5.0
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.3
